script/requirements.txt
torch==1.7
torchvision==0.8.2
base64
protobuf
grpcio
grpcio-tools
